30 hours in, and we have a tileset and working missions/promotion.

+++ src/components/renderable.hpp
@@ -3,9 +3,9 @@
 #include <rltk.hpp>
 
 struct renderable_t {
-    uint8_t glyph = '@';
+    uint16_t glyph = '@';
     rltk::color_t color = rltk::colors::YELLOW;
 
     renderable_t() {}
-    renderable_t(const uint8_t &g, const rltk::color_t &c) : glyph(g), color(c) {}
+    renderable_t(const uint16_t &g, const rltk::color_t &c) : glyph(g), color(c) {}
 };
